+++ examples/complete/data.tf
@@ -1,0 +1,53 @@
+data "aws_caller_identity" "current" {}
+
+data "aws_iam_policy_document" "sns_topic_policy" {
+  policy_id = "example_policy_ID"
+
+  statement {
+    actions = [
+      "SNS:Subscribe",
+      "SNS:SetTopicAttributes",
+      "SNS:RemovePermission",
+      "SNS:Receive",
+      "SNS:Publish",
+      "SNS:ListSubscriptionsByTopic",
+      "SNS:GetTopicAttributes",
+      "SNS:DeleteTopic",
+      "SNS:AddPermission",
+    ]
+
+    condition {
+      test     = "StringEquals"
+      variable = "AWS:SourceOwner"
+
+      values = [
+        local.account_id,
+      ]
+    }
+
+    effect = "Allow"
+
+    principals {
+      type        = "AWS"
+      identifiers = ["*"]
+    }
+
+    resources = [
+      module.complete_sns_topic.arn,
+    ]
+
+    sid = "example_statement_ID"
+  }
+}
+
+data "aws_iam_policy_document" "sns_assume_role_policy" {
+  statement {
+    sid     = "snsSTSAssumeRole"
+    actions = ["sts:AssumeRole"]
+
+    principals {
+      type        = "Service"
+      identifiers = ["sns.amazonaws.com"]
+    }
+  }
+}