+++ CHANGELOG.md
@@ -5,9 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Feature: Expand the complete example
 - Add an example that uses an external CMK KMS
-- Add subscription examples on the complete for multiple endpoints, this will be also on the complete
+- showcase subscription_role_arn to publish to Kinesis Data Firehose delivery stream.
+- Add subscription examples on the complete for other protocols (lambda, HTTP, platform application endpoint, SMS, kinesis data firehorse ).
+
+## [1.1.2] - 2023-10-23
+### Changes
+- fix: sns topic policy
+- added example sqs queue to used as dead letter queue arn
+- added example filter policy for sns topic subscription
+- added example roles to showcase success feedback arns
+- added email and https protocol examples for sns topic subscription
 
 ## [1.1.1] - 2022-07-19
 ### Changes
@@ -40,8 +48,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial commit
 - sns module with an example
 
-[Unreleased]: https://github.com/boldlink/terraform-aws-sns/compare/1.1.0...HEAD
+[Unreleased]: https://github.com/boldlink/terraform-aws-sns/compare/1.1.2...HEAD
 
+[1.1.2]: https://github.com/boldlink/terraform-aws-sns/releases/tag/1.1.2
 [1.1.1]: https://github.com/boldlink/terraform-aws-sns/releases/tag/1.1.1
 [1.1.0]: https://github.com/boldlink/terraform-aws-sns/releases/tag/1.1.0
 [1.0.2]: https://github.com/boldlink/terraform-aws-sns/releases/tag/1.0.2